C0: extract Matcher to own the ladders and pure decision helpers
Stands up Matcher holding the working/stop PriceLadders, and moves
TryComputeAuctionPrice, HasSufficientLiquidity, and the self-match
verdict (IsSelfMatch/TryGetSelfMatchInstruction) onto it. Match's
control flow (the loop, event emission, stop-checking) stays on
InMemoryOrderBook, now delegating ladder access and these decisions to
_matcher. No behavior change - proves the seam ahead of moving the
Match loop itself.
